backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.1
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.104.1
google-auth==2.25.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.1.53
langchain-openai==0.0.5
langsmith==0.1.147
multidict==6.7.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.4
packaging==23.2
postgrest==2.24.0
propcache==0.4.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
realtime==2.24.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.27.0
storage3==2.24.0
StrEnum==0.4.15
stripe==7.8.0
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
tenacity==8.5.0
tiktoken==0.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
